% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatTextcircle.R
\name{compute_panel_textcircle}
\alias{compute_panel_textcircle}
\title{Compute textcircle}
\usage{
compute_panel_textcircle(data, scales, r = 3, x0 = 0, y0 = 0)
}
\arguments{
\item{data}{data}

\item{scales}{scales}

\item{r}{radius of inner circle}

\item{x0}{x coordinate of inner circle}

\item{y0}{y coordinate of inner circle}
}
\description{
Compute textcircle
}
