% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_textcircle.R
\name{geom_textcircle}
\alias{geom_textcircle}
\title{Geom function for textcircle}
\usage{
geom_textcircle(
  geom = ggplot2::GeomText,
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  hjust = 1,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{geom}{geom}

\item{mapping}{mapping}

\item{data}{data}

\item{position}{position}

\item{na.rm}{na.rm}

\item{hjust}{alignment of text}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{inherited}
}
\description{
Geom function for textcircle
}
