% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_textcircle.R
\name{stat_textcircle}
\alias{stat_textcircle}
\title{Stat function for textcircle}
\usage{
stat_textcircle(
  geom = ggplot2::GeomText,
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{geom}{geom}

\item{mapping}{mapping}

\item{data}{data}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{inherited}
}
\description{
Stat function for textcircle
}
